backend/requirements.txt
fastapi
uvicorn
python-jose
passlib
python-multipart
pydantic
python-dotenv
requests
